Update
Fixed position of turret to straight up.
Use WASD to move turret around.
In order to shoot multiple cannon balls we need to have the cannon ball
be separate from the base. Which is initially done, but then when i try
to move it (the base), the cannon resets to the original pos. Animation
needs to be changed a bit. You can see what I attempted, but no real
progress with it.